category_service/pyproject.toml
[tool.poetry]
name = "category-service"
version = "0.1.0"
description = ""
authors = ["Klukvas <pavlenkoandrey56@gmail.com>"]
readme = "README.md"
package-mode = false 


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.115.12"
uvicorn = "0.34.0"
SQLAlchemy = "2.0.40"
psycopg2-binary = "2.9.10"
httpx = "0.28.1"
pydantic = {extras = ["email"], version = "^2.11.3"}
pydantic-settings = "2.8.1"
python-jose = "3.4.0"
python-dotenv = "1.1.0"
alembic = "1.15.2"
anyio = "4.9.0"
certifi = "2025.1.31"
click = "8.1.8"
ecdsa = "0.19.1"
h11 = "0.14.0"
httpcore = "1.0.7"
idna = "3.10"
Mako = "1.3.10"
MarkupSafe = "3.0.2"
packaging = "24.2"
pyasn1 = "0.4.8"
rsa = "4.9"
six = "1.17.0"
sniffio = "1.3.1"
starlette = "0.46.1"
typing-inspection = "0.4.0"
typing_extensions = "4.13.2"
tzdata = "2025.2"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.5"
Faker = "37.1.0"
iniconfig = "2.1.0"
pluggy = "1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"